requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-groq
chromadb==0.4.22
openai==1.7.2
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers==2.2.2
playwright==1.41.0
